+++ workspace/assignments/03-lidar-odometry-advanced/src/lidar_localization/include/lidar_localization/scan_map_registration/scan_map_registration.hpp
@@ -70,8 +70,8 @@ class ScanMapRegistration {
     // odometry:
     //
     struct Pose {
-      Eigen::Quaternionf q = Eigen::Quaternionf::Identity();
-      Eigen::Vector3f t = Eigen::Vector3f::Zero();
+      Eigen::Quaterniond q = Eigen::Quaterniond::Identity();
+      Eigen::Vector3d t = Eigen::Vector3d::Zero();
     };
 
     struct {
